Added render pass type parameter
Render passes now have a type parameter, which in turn required
type parameters on several other types.  The purpose of this is
to prevent incompatible render passes from being used accidentally.
There is also now a RenderPassFactory trait, which can be implemented
to produce objects that create render passes with the same type
parameter.  So long as the RenderPassFactory is implemented correctly,
the produced render passes should all be compatible with each other.
Any attempt to use render passes produced by a different factory
(likely to be incompatible) should result in a compile error.