Split Presenter::render() into two functions
As it turns out, there was a reason I previously had these two
separated.  Trying to write a command buffer while holding a
mutable reference to the Presenter is a borrow checker nightmare.
There might be something really, really stupid I can do to get
around it, but this solution is far more elegant.

Just don't drop a Frame object on the floor, please.