Added TemporaryCommandBufferSet
This is a mechanism for retaining references to one-shot command
buffers until they have finished rendering.  The idea is that you
have an Option<Rc<CommandBuffer>> for each swapchain image, and
after waiting for that image's fence, you set that Option to None.
Thus, the command buffer will not be destroyed until the GPU is
done using it.