Yet another rework of Presenter
The inflight fences are gone; those are now the user's responsibility.
The motivation for this change is the realization that these fences
pull double duty: signalling that the swapchain image can be reused
and signalling that the command buffer is no longer in use.  Since
managing command buffers is the user's responsibility, the fences
used by the command buffers also need to be the user's responsibility.

In addition, present() now supports multiple command buffers, complete
with semaphores to enforce dependencies between them.  All semaphores
except for the image-available and render-finished semaphores will be
furnished by the user.  It is up to the user to ensure that the correct
command buffers wait for image-available and signal render-finished.
It is also up to the user to provide a render-finished fence.  Calling
present() hands references to the command buffers, semaphores, and
fences over to the Presenter, which stores them until the swapchain
image that they were used with comes up again.  At that point, it will
wait for the render-finished fence, then remove its references.  If
the command buffers are one-shot and have no references anywhere else,
they will be destroyed at this point, along with their semaphores
and fences if those are not referred to elsewhere.

I'm still not 100% certain that this is the correct design when using
one-shot command buffers and non-transient fences.  If the fences
persist and get used with new command buffers, the Presenter may end
up waiting for a command buffer that it doesn't really need to wait
for.

Maybe the solution there is simply "don't do that."