More presentation API tweaks
As it turns out, I had to preserve the render-finished fences on
the user side; they're needed to keep secondary command buffers
from being executed concurrently.  Since I needed to keep those
fences in a PerFrameSet, I needed to know the frame ID in order
to retrieve the correct fence, but I couldn't know the frame ID
until calling acquire_image(), so having acquire_image() take the
fence as a parameter didn't work so well.

I guess the "needless" waiting isn't so needless after all.  While
there may not be a need to wait for the swapchain image due to
fence reuse (e.g. fence 1 was used with swapchain image A last time
and is now being used with swapchain image C), the wait is still
needed to avoid concurrent command buffer execution.