Just wait for the fence from the current submission set
The swapchain system won't give us a swapchain image unless that
image has been presented (more or less).  So, we don't have to
worry about the swapchain image being in use in present().  We
do, however, have to worry about the command buffer being in use,
so we wait for the render-finished fence and then reset it.

The wait-fence has been removed; it's now redundant.  Instead of
waiting for the fence in the old submission set, we now just set
it to None to release its resources, since by this point we know
that they're no longer in use.