Added HasHandle trait
This trait allows the user to retrieve the underlying Vulkan handle
for various objects.  This is only for debugging at this point;
I may eventually remove it and add a to_vk() function for all
relevant types instead.  That would return the ash::vk::Whatever
object that the type holds instead of returning a u64 handle.