User can now enable Vulkan features
The user can now give a DeviceBuilder a pointer to a chain of
Features structures (e.g. the atomic float features structure)
to enable whatever Vulkan features they need.  As part of this
change, non-uniform access is no longer enabled by default, since
it is not required for any ashbrick functionality.  Its presence
was a holdover from the pre-independence days.