Massive command/queue rework
This is a result of me gaining a much better understanding of how
command buffers, pools, queue families, and queues work.  Some
things are a little more verbose (e.g. image transitions and buffer
copies require the user to provide a command pool and a queue for
the generated command buffers to be allocated from and executed
through, respectively), but more control is now in the user's hands.