Removed CommandBufferSubmissionSet::render_finished_fence
As it turns out, there are things you want to avoid doing until rendering
is complete that aren't part of a command buffer (e.g. writing to buffers
that are host-visible and coherent).  Thus, waiting for rendering to be
complete is something that the calling code needs to do well in advance
of calling Presenter::present().  The Presenter itself doesn't need to
know anything about rendering being finished; its only concern is with
swapchain images being in use or not.